Move programs to programs folder
Moved all files related to the programs sub-menu to its own folder.
