+++ wondermom/src/App.js
@@ -12,9 +12,9 @@ import Backdrop from './components/Backdrop/Backdrop';
 import Home from './pages/Home';
 import About from './pages/About';
 import Programs from './pages/Programs';
-import DayOne from "./pages/DayOne";
-import SixWeekProgram from "./pages/SixWeekProgram";
-import WeekOne from "./pages/WeekOne";
+import DayOne from "./pages/Programs/DayOne";
+import SixWeekProgram from "./pages/Programs/SixWeekProgram";
+import WeekOne from "./pages/Programs/WeekOne";
 import Nutrition from './pages/Nutritionpages/Nutrition';
 import Breakfast from './pages/Nutritionpages/Breakfast';
 import Lunch from './pages/Nutritionpages/Lunch';
